Update instructions for Pop_OS prior to 20.04
Pop_OS 20.04 uses flatpak remotes in user mode by default. In order to simplify the transition for users upgrading from 19.10 or earlier, the instructions should have users on these versions adding Flathub in user-mode instead of system-mode, because otherwise the remote has to be removed and re-added in order to continue using Flathub from Pop_Shop in 20.04.